Enhanced design
