+++ src/app.html
@@ -1,15 +1,15 @@
 <!doctype html>
 <html lang="en">
-	<head>
-		<meta charset="utf-8" />
-		<link rel="icon" href="%sveltekit.assets%/money-24.png" />
-		<link rel="manifest" href="/manifest.json" />
-		<meta name="theme-color" content="#4075a6" />
-		<meta name="viewport" content="width=device-width, initial-scale=1" />
+  <head>
+    <meta charset="utf-8" />
+    <link rel="icon" href="%sveltekit.assets%/money-24.png" />
+    <link rel="manifest" href="/manifest.json" />
+    <meta name="theme-color" content="#053661" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
 
-		%sveltekit.head%
-	</head>
-	<body data-sveltekit-preload-data="hover">
-		<div style="display: contents">%sveltekit.body%</div>
-	</body>
+    %sveltekit.head%
+  </head>
+  <body data-sveltekit-preload-data="hover">
+    <div style="display: contents">%sveltekit.body%</div>
+  </body>
 </html>
