+++ src/lib/debugger.ts
@@ -1,6 +1,6 @@
 import { logs } from './store/logs';
 
 export function debug(message: string) {
-  logs.update((l) => [...l, `c: ${message}`]);
-  console.log(`c: ${message}`);
+  logs.update((l) => [...l, `ct: ${message}`]);
+  console.log(` c: ${message}`);
 }
