+++ src/routes/styles.css
@@ -4,11 +4,10 @@
   --font-body: Arial, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu,
     Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
   --font-mono: 'Fira Mono', monospace;
-  --color-bg-1: hsl(208, 90%, 20%);
-  --color-bg-2: hsl(208, 86%, 14%);
-  --color-theme-1: #ff3e00;
-  --color-theme-2: #4075a6;
-  --color-text: rgba(0, 0, 0, 0.7);
+  --color-bg-1: #053661;
+  --color-bg-2: #052642;
+  --color-bg-3: #4075a6;
+  --color-text: #000000b3;
   --color-text-2: #5391e0;
   --column-width: 42rem;
   --column-margin-top: 4rem;
@@ -22,10 +21,4 @@ body {
   background-attachment: fixed;
   background-color: var(--color-bg-1);
   background-size: 100vw 100vh;
-  background-image: radial-gradient(
-      2 50% 50% at 50% 50%,
-      rgba(255, 255, 255, 0.75) 0%,
-      rgba(255, 255, 255, 0) 100%
-    ),
-    linear-gradient(180deg, var(--color-bg-0) 0%, var(--color-bg-1) 15%, var(--color-bg-2) 50%);
 }
